Restructured deid files into multiple files
