+++ dlp/snippets/Deidentify/deidentify_redact_test.py
@@ -1,0 +1,30 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the 'License');
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an 'AS IS' BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+import os
+
+import deidentify_redact as deid
+
+import pytest
+
+
+HARMFUL_STRING = "My SSN is 372819127"
+GCLOUD_PROJECT = os.getenv("GOOGLE_CLOUD_PROJECT")
+
+
+def test_deidentify_with_redact(capsys: pytest.CaptureFixture) -> None:
+    deid.deidentify_with_redact(
+        GCLOUD_PROJECT, HARMFUL_STRING + "!", ["US_SOCIAL_SECURITY_NUMBER"]
+    )
+    out, _ = capsys.readouterr()
+    assert "My SSN is !" in out
