+++ dlp/snippets/Deidentify/reidentify_text_fpe.py
@@ -1,0 +1,141 @@
+# Copyright 2023 Google LLC
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+"""Uses of the Data Loss Prevention API for deidentifying sensitive data."""
+
+from __future__ import annotations
+
+import argparse
+
+import base64
+
+import google.cloud.dlp
+
+# [START dlp_reidentify_text_fpe]
+
+
+def reidentify_text_with_fpe(
+    project: str,
+    input_str: str,
+    key_name: str = None,
+    wrapped_key: str = None,
+) -> None:
+    """
+    Uses the Data Loss Prevention API to re-identify sensitive data in a
+    string using Format Preserving Encryption (FPE).
+    Args:
+        project: The Google Cloud project id to use as a parent resource.
+        input_str: The string to re-identify (will be treated as text).
+        key_name: The name of the Cloud KMS key used to encrypt ('wrap') the
+            AES-256 key. Example:
+            key_name = 'projects/YOUR_GCLOUD_PROJECT/locations/YOUR_LOCATION/
+            keyRings/YOUR_KEYRING_NAME/cryptoKeys/YOUR_KEY_NAME'
+        wrapped_key: The encrypted ('wrapped') AES-256 key to use. This key
+            should be encrypted using the Cloud KMS key specified by key_name.
+    """
+
+    # Instantiate a client.
+    dlp = google.cloud.dlp_v2.DlpServiceClient()
+
+    # The wrapped key is base64-encoded, but the library expects a binary
+    # string, so decode it here.
+    wrapped_key = base64.b64decode(wrapped_key)
+
+    # Specify the type of info the inspection will re-identify.
+    # This must use the same custom infoType that was used as a
+    # surrogate during the initial encryption.
+    surrogate_info_type = {"name": "PHONE_NUMBER"}
+
+    # Construct FPE configuration dictionary
+    crypto_replace_ffx_fpe_config = {
+        "crypto_key": {
+            "kms_wrapped": {"wrapped_key": wrapped_key, "crypto_key_name": key_name}
+        },
+        "common_alphabet": "NUMERIC",
+        "surrogate_info_type": surrogate_info_type,
+    }
+
+    # Construct re-identify configuration dictionary
+    reidentify_config = {
+        "info_type_transformations": {
+            "transformations": [
+                {
+                    "primitive_transformation": {
+                        "crypto_replace_ffx_fpe_config": crypto_replace_ffx_fpe_config,
+                    },
+                    "info_types": [surrogate_info_type],
+                }
+            ]
+        }
+    }
+
+    # Construct inspect configuration dictionary
+    inspect_config = {
+        "custom_info_types": [{"info_type": surrogate_info_type, "surrogate_type": {}}]
+    }
+
+    # Construct the `item`.
+    item = {"value": input_str}
+
+    # Convert the project id into a full resource id.
+    parent = f"projects/{project}/locations/global"
+
+    # Call the API.
+    response = dlp.reidentify_content(
+        request={
+            "parent": parent,
+            "reidentify_config": reidentify_config,
+            "inspect_config": inspect_config,
+            "item": item,
+        }
+    )
+
+    # Print results
+    print(f"Text after re-identification: {response.item.value}")
+
+
+# [END dlp_reidentify_text_fpe]
+
+if __name__ == "__main__":
+    parser = argparse.ArgumentParser(description=__doc__)
+
+    parser.add_argument(
+        "project",
+        help="The Google Cloud project id to use as a parent resource.",
+    )
+    parser.add_argument(
+        "item",
+        help="The string to deidentify. " "Example: string = 'My SSN is 372819127'",
+    )
+    parser.add_argument(
+        "key_name",
+        help="The name of the Cloud KMS key used to encrypt ('wrap') the "
+             "AES-256 key. Example: "
+             "key_name = 'projects/YOUR_GCLOUD_PROJECT/locations/YOUR_LOCATION/"
+             "keyRings/YOUR_KEYRING_NAME/cryptoKeys/YOUR_KEY_NAME'",
+    )
+    parser.add_argument(
+        "wrapped_key",
+        help="The encrypted ('wrapped') AES-256 key to use. This key should "
+             "be encrypted using the Cloud KMS key specified by key_name.",
+    )
+
+    args = parser.parse_args()
+
+    reidentify_text_with_fpe(
+        args.project,
+        args.item,
+        wrapped_key=args.wrapped_key,
+        key_name=args.key_name,
+    )
